+++ src/licenses/infrastructure/repositories/LicenseJSDataverseRepository.ts
@@ -1,0 +1,9 @@
+import { LicenseRepository } from '../../domain/repositories/LicenseRepository'
+import { License } from '../../domain/models/License'
+import { getAvailableStandardLicenses } from '@iqss/dataverse-client-javascript'
+
+export class LicenseJSDataverseRepository implements LicenseRepository {
+  async getAvailableStandardLicenses(): Promise<License[]> {
+    return (await getAvailableStandardLicenses.execute()) as License[]
+  }
+}